blender-worker/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pillow==11.0.0
numpy==2.1.3
requests==2.32.3
